Relocate stats class to ML package
